pyproject.toml
[project]
name = "hacks-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
